se eliminan comentarios y se agrega la clase Character como padre de NonPlayableCharacter y Personaje
